GitHub Action: improved maintenance script performance for better efficiency

+++ scripts/maintenance.py
@@ -25,18 +25,22 @@ print('-------------------')
 tz = pytz.timezone('Europe/Madrid')
 dt = datetime.now(tz)
 
-print(f'Expired links at {dt.strftime("%d/%m/%Y %H:%M:%S")}:')
+print(f'Expired keys at {dt.strftime("%d/%m/%Y %H:%M:%S")}:')
+cursor = conn.execute("SELECT apiKey, expires_at FROM keys WHERE expires_at < DATETIME('now');")
+rows = cursor.fetchall()
 
-cursor = conn.execute("SELECT apiKey FROM keys WHERE expires_at < DATETIME('now');")
-for key in cursor.fetchall():
-    print(f'- {key[0]}')
+print(f'Expired keys found: {len(rows)}')
+for row in rows:
+    print(f'- {row[0]}')
+
+if not rows:
+    print('No expired keys found!')
+    exit()
 
 print(f'Deleting ...')
-
 conn.execute("PRAGMA foreign_keys = ON;")
 cursor = conn.execute("DELETE FROM keys WHERE expires_at < DATETIME('now');")
 conn.commit()
-
 conn.sync()
 
 print(f'Deleted {cursor.rowcount} expired keys!')
